+++ app/jobs/users/export_data_job.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+class Users::ExportDataJob < ApplicationJob
+  queue_as :exports
+
+  def perform(user_id)
+    user = User.find(user_id)
+
+    Users::ExportData.new(user).export
+  end
+end