+++ spec/models/import_spec.rb
@@ -8,6 +8,11 @@ RSpec.describe Import, type: :model do
     it { is_expected.to belong_to(:user) }
   end
 
+  describe 'validations' do
+    it { is_expected.to validate_presence_of(:name) }
+    it { is_expected.to validate_uniqueness_of(:name).scoped_to(:user_id) }
+  end
+
   describe 'enums' do
     it do
       is_expected.to define_enum_for(:source).with_values(